// src/app/components/merchandising/purchase-order/Booking-Instruction/model/booking-instruction.ts
export class BookingInstruction {
  bookingInstructionId: number;
  fileStatusId:number;
  brand:string;
  spl: string;
  iman:string;
  seasonId: number;
  seasonName: string;
  styleNo: string;
  styleId: number;
  modelNo: string;
  color:string;
  itemDescription:string;
  globalSelection: number;
  europeSelection: number;
  totalReceived: number;
  totalBooking: number;
  totalProduction: number;
  totalStock: number;

  euSs: number;
  euAw: number;
  europeElasticityPlus: number;
  europeElasticityMinus: number;
  europeSupply: string;
  europeReceived: number;
  europeBooking: number;

  morocco: number;
  moroccoElasticityPlus: number;
  moroccoElasticityMinus: number;
  moroccoSupply: string;
  moroccoReceived: number;
  moroccoBooking: number;

  turkey: number;
  turkeyElasticityPlus: number;
  turkeyElasticityMinus: number;
  turkeySupply: string;
  turkeyReceived: number;
  turkeyBooking: number;

  russia: number;
  russiaElasticityPlus: number;
  russiaElasticityMinus: number;
  russiaSupply: string;
  russiaReceived: number;
  russiaBooking: number;

  india: number;
  indiaElasticityPlus: number;
  indiaElasticityMinus: number;
  indiaSupply: string;
  indiaReceived: number;
  indiaBooking: number;

  brazil: number;
  brazilElasticityPlus: number;
  brazilElasticityMinus: number;
  brazilSupply: string;
  brazilReceived: number;
  brazilBooking: number;

  asia: number;
  asiaElasticityPlus: number;
  asiaElasticityMinus: number;
  asiaSupply: string;
  asiaReceived: number;
  asiaBooking: number;

  america: number;
  americaElasticityPlus: number;
  americaElasticityMinus: number;
  americaSupply: string;
  americaReceived: number;
  americaBooking: number;

  china: number;
  chinaElasticityPlus: number;
  chinaElasticityMinus: number;
  chinaSupply: string;
  chinaReceived: number;
  chinaBooking: number;

  seasonNamePO: string;
  styleNoPO: string;
  model: string;
  receivedQty: number;
  deliveryPlaceId: number;
  deliveryPlaceName: string;
  countryName: string;
  delDate: Date;
  unitPrice: number;

  canada: number;
  canadaElasticityPlus: number;
  canadaElasticityMinus: number;
  canadaSupply: string;
  canadaReceived: number;
  canadaBooking: number;

  indianOcean: number;
  indianOceanElasticityPlus: number;
  indianOceanElasticityMinus: number;
  indianOceanSupply: string;
  indianOceanReceived: number;
  indianOceanBooking: number;

  thailand: number;
  thailandElasticityPlus: number;
  thailandElasticityMinus: number;
  thailandSupply: string;
  thailandReceived: number;
  thailandBooking: number;

  uk: number;
  ukElasticityPlus: number;
  ukElasticityMinus: number;
  ukSupply: string;
  ukReceived: number;
  ukBooking: number;

  brandId: number;
  brandName: string;
  stockComments: string;
  remarks: string;
  updateBy: string;
  buyerId: number;
  seasonYearId: number;
  updateStatus: number;
  fileTrackNo:number;
}